requirements.txt
websocket-client
psutil